+++ CHANGELOG.md
@@ -14,5 +14,6 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip.
 
 ## [0.2.x](https://github.com/snakemake/snakedeploy/tree/master) (0.0.x)
+ - adding option to template with GitHub (if GitHub token is present) (0.0.11)
  - beta release with working clone (0.0.1)
  - skeleton release (0.0.0)